fix: separate GitHub import and connect flows
