+++ frontend/app/(main)/projects/[projectId]/data/components/github-integration/GithubIntegrationPanel.tsx
@@ -2,13 +2,13 @@
 
 /**
  * GithubIntegrationPanel — the project-level GitHub binding UI,
- * mounted inline inside the "New access → GitHub" flow.
+ * mounted inline inside the "Connect GitHub" flow.
  *
  * This is the same flow the now-deleted ``/projects/{id}/integrations``
  * top-level page used to drive; that route was removed in favour of
- * consolidating the GitHub binding under the same scope-level connector
- * picker that hosts Notion / Gmail / Google Docs / etc., per the
- * 2026-05-10 UX consolidation.
+ * consolidating the GitHub binding under the same durable-source picker
+ * that hosts Notion / Gmail / Google Docs / etc., per the 2026-05-10
+ * UX consolidation.
  *
  * The binding itself is still **project-level** — it's not actually
  * scoped to whichever folder the user opened the picker from. We
